fix lint by moving cheats_test into runner_test
